pyproject.toml
[tool.poetry]
name = "radar-iot"
version = "0.1.0"
description = ""
authors = ["eric li"]
readme = "README.md"
packages = [{include = "radar_iot"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.109.2"}
paho-mqtt = "^2.0.0"
ksoc-connection = {git = "https://github.com/lianyun0502/KSOC_Connection.git"}
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlalchemy = "^2.0.27"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
